+++ app/models/user_site_association.rb
@@ -7,6 +7,7 @@
 #  user_id    :integer
 #  created_at :datetime         not null
 #  updated_at :datetime         not null
+#  role       :integer          default(3)
 #
 
 class UserSiteAssociation < ApplicationRecord
@@ -15,4 +16,18 @@ class UserSiteAssociation < ApplicationRecord
 
   scope :manager, -> { where(role: UserType::MANAGER) }
   scope :publisher, -> { where(role: UserType::PUBLISHER) }
+
+  validates :user, uniqueness: {scope: :site}
+
+  attr_accessor :selected # virtual attribute used for processing the form
+
+  def <=> another
+    if user.try(:name) < another.user.try(:name)
+      -1
+    elsif user.try(:name) > another.user.try(:name)
+      1
+    else
+      0
+    end
+  end
 end